onedev/requirements.txt
Django==4.1.5
psycopg2-binary==2.9.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
Pillow==9.4.0
djangorestframework-simplejwt==5.2.2
